requirements.txt
slack_bolt
fastapi
uvicorn
psycopg2-binary
pandas
python-dotenv
html5lib      # if you ever parse more HTML data
requests      # if making external API calls
